Refactor MappingQ/Q1 into a common base class.
This patch separates out the common functionality for generic polynomial
mappings into a base class. It contains most of what used to be in MappingQ1
but really was not specific to p=1 mappings.
