+++ include/deal.II/fe/mapping_q1.h
@@ -23,7 +23,7 @@
 #include <deal.II/base/qprojector.h>
 #include <deal.II/grid/tria_iterator.h>
 #include <deal.II/dofs/dof_accessor.h>
-#include <deal.II/fe/mapping.h>
+#include <deal.II/fe/mapping_q_generic.h>
 
 #include <cmath>
 
@@ -37,7 +37,8 @@ DEAL_II_NAMESPACE_OPEN
  * Mapping of the reference to cell to a general
  * quadrilateral/hexahedra by $d$-linear shape functions.
  *
- * This function maps the reference (unit) cell to a general grid cell with
+ * This mapping implemented by this class maps the reference (unit) cell
+ * to a general grid cell with
  * straight lines in $d$ dimensions. (Note, however, that in 3D the
  * <i>faces</i> of a general, trilinearly mapped cell may be curved, even if the
  * edges are not). This is the standard mapping used for polyhedral domains. It
@@ -51,7 +52,7 @@ DEAL_II_NAMESPACE_OPEN
  * @author Guido Kanschat, 2000, 2001; Ralf Hartmann, 2000, 2001, 2005
  */
 template <int dim, int spacedim=dim>
-class MappingQ1 : public Mapping<dim,spacedim>
+class MappingQ1 : public MappingQGeneric<dim,spacedim>
 {
 public:
   /**
@@ -62,12 +63,6 @@ public:
   // for documentation, see the Mapping base class
   virtual
   Mapping<dim,spacedim> *clone () const;
-
-  /**
-   * Always returns @p true because MappingQ1 preserves vertex locations.
-   */
-  virtual
-  bool preserves_vertex_locations () const;
 
   /**
    * @name Mapping points between reference and real cells
@@ -90,54 +85,6 @@ public:
    * @}
    */
 
-  /**
-   * @name Functions to transform tensors from reference to real coordinates
-   * @{
-   */
-
-  // for documentation, see the Mapping base class
-  virtual
-  void
-  transform (const VectorSlice<const std::vector<Tensor<1,dim> > >   input,
-             const MappingType                                       type,
-             const typename Mapping<dim,spacedim>::InternalDataBase &internal,
-             VectorSlice<std::vector<Tensor<1,spacedim> > >          output) const;
-
-  // for documentation, see the Mapping base class
-  virtual
-  void
-  transform (const VectorSlice<const std::vector<DerivativeForm<1, dim,spacedim> > > input,
-             const MappingType                                                       type,
-             const typename Mapping<dim,spacedim>::InternalDataBase                 &internal,
-             VectorSlice<std::vector<Tensor<2,spacedim> > >                          output) const;
-
-  // for documentation, see the Mapping base class
-  virtual
-  void
-  transform (const VectorSlice<const std::vector<Tensor<2, dim> > >  input,
-             const MappingType                                       type,
-             const typename Mapping<dim,spacedim>::InternalDataBase &internal,
-             VectorSlice<std::vector<Tensor<2,spacedim> > >          output) const;
-
-  // for documentation, see the Mapping base class
-  virtual
-  void
-  transform (const VectorSlice<const std::vector< DerivativeForm<2, dim, spacedim> > > input,
-             const MappingType                                                         type,
-             const typename Mapping<dim,spacedim>::InternalDataBase                   &internal,
-             VectorSlice<std::vector<Tensor<3,spacedim> > >                            output) const;
-
-  // for documentation, see the Mapping base class
-  virtual
-  void
-  transform (const VectorSlice<const std::vector<Tensor<3, dim> > >  input,
-             const MappingType                                       type,
-             const typename Mapping<dim,spacedim>::InternalDataBase &internal,
-             VectorSlice<std::vector<Tensor<3,spacedim> > >          output) const;
-
-  /**
-   * @}
-   */
 
   /**
    * @name Interface with FEValues
@@ -146,298 +93,13 @@ public:
 
 public:
   /**
-   * Storage for internal data of d-linear mappings. See Mapping::InternalDataBase
-   * for an extensive description.
-   *
-   * For the current class MappingQ1, the InternalData class store
-   * data that is computed once when the object is created
-   * (in get_data()) as well as data the class wants to store from between
-   * the call to fill_fe_values(), fill_fe_face_values(), or
-   * fill_fe_subface_values() until possible later calls from the finite
-   * element to functions such as transform(). The latter class of
-   * member variables are marked as 'mutable'.
+   * Use the InternalData class of the base class without modification
+   * and additions.
    */
-  class InternalData : public Mapping<dim,spacedim>::InternalDataBase
-  {
-  public:
-    /**
-     * Constructor. The argument denotes the polynomial degree of
-     * the mapping to which this object will correspond.
-     */
-    InternalData(const unsigned int polynomial_degree);
-
-    /**
-     * Initialize the object's member variables related to cell data
-     * based on the given arguments.
-     *
-     * The function also calls compute_shape_function_values() to
-     * actually set the member variables related to the values and
-     * derivatives of the mapping shape functions.
-     */
-    void
-    initialize (const UpdateFlags      update_flags,
-                const Quadrature<dim> &quadrature,
-                const unsigned int     n_original_q_points);
-
-    /**
-     * Initialize the object's member variables related to cell and
-     * face data based on the given arguments. In order to initialize
-     * cell data, this function calls initialize().
-     */
-    void
-    initialize_face (const UpdateFlags      update_flags,
-                     const Quadrature<dim> &quadrature,
-                     const unsigned int     n_original_q_points);
-
-    /**
-     * Compute the values and/or derivatives of the shape functions
-     * used for the mapping.
-     *
-     * Which values, derivatives, or higher order derivatives are
-     * computed is determined by which of the member arrays have
-     * nonzero sizes. They are typically set to their appropriate
-     * sizes by the initialize() and initialize_face() functions,
-     * which indeed call this function internally. However, it is
-     * possible (and at times useful) to do the resizing by hand and
-     * then call this function directly. An example is in a Newton
-     * iteration where we update the location of a quadrature point
-     * (e.g., in MappingQ::transform_real_to_uni_cell()) and need to
-     * re-compute the mapping and its derivatives at this location,
-     * but have already sized all internal arrays correctly.
-     */
-    void compute_shape_function_values (const std::vector<Point<dim> > &unit_points);
-
-
-    /**
-     * Shape function at quadrature point. Shape functions are in tensor
-     * product order, so vertices must be reordered to obtain transformation.
-     */
-    const double &shape (const unsigned int qpoint,
-                         const unsigned int shape_nr) const;
-
-    /**
-     * Shape function at quadrature point. See above.
-     */
-    double &shape (const unsigned int qpoint,
-                   const unsigned int shape_nr);
-
-    /**
-     * Gradient of shape function in quadrature point. See above.
-     */
-    const Tensor<1,dim> &derivative (const unsigned int qpoint,
-                                     const unsigned int shape_nr) const;
-
-    /**
-     * Gradient of shape function in quadrature point. See above.
-     */
-    Tensor<1,dim> &derivative (const unsigned int qpoint,
-                               const unsigned int shape_nr);
-
-    /**
-     * Second derivative of shape function in quadrature point. See above.
-     */
-    const Tensor<2,dim> &second_derivative (const unsigned int qpoint,
-                                            const unsigned int shape_nr) const;
-
-    /**
-     * Second derivative of shape function in quadrature point. See above.
-     */
-    Tensor<2,dim> &second_derivative (const unsigned int qpoint,
-                                      const unsigned int shape_nr);
-
-    /**
-     * third derivative of shape function in quadrature point. See above.
-     */
-    const Tensor<3,dim> &third_derivative (const unsigned int qpoint,
-                                           const unsigned int shape_nr) const;
-
-    /**
-     * third derivative of shape function in quadrature point. See above.
-     */
-    Tensor<3,dim> &third_derivative (const unsigned int qpoint,
-                                     const unsigned int shape_nr);
-
-    /**
-     * fourth derivative of shape function in quadrature point. See above.
-     */
-    const Tensor<4,dim> &fourth_derivative (const unsigned int qpoint,
-                                            const unsigned int shape_nr) const;
-
-    /**
-     * fourth derivative of shape function in quadrature point. See above.
-     */
-    Tensor<4,dim> &fourth_derivative (const unsigned int qpoint,
-                                      const unsigned int shape_nr);
-
-    /**
-     * Return an estimate (in bytes) or the memory consumption of this object.
-     */
-    virtual std::size_t memory_consumption () const;
-
-    /**
-     * Values of shape functions. Access by function @p shape.
-     *
-     * Computed once.
-     */
-    std::vector<double> shape_values;
-
-    /**
-     * Values of shape function derivatives. Access by function @p derivative.
-     *
-     * Computed once.
-     */
-    std::vector<Tensor<1,dim> > shape_derivatives;
-
-    /**
-     * Values of shape function second derivatives. Access by function @p
-     * second_derivative.
-     *
-     * Computed once.
-     */
-    std::vector<Tensor<2,dim> > shape_second_derivatives;
-
-    /**
-     * Values of shape function third derivatives. Access by function @p
-     * second_derivative.
-     *
-     * Computed once.
-     */
-    std::vector<Tensor<3,dim> > shape_third_derivatives;
-
-    /**
-     * Values of shape function fourth derivatives. Access by function @p
-     * second_derivative.
-     *
-     * Computed once.
-     */
-    std::vector<Tensor<4,dim> > shape_fourth_derivatives;
-
-    /**
-     * Unit tangential vectors. Used for the computation of boundary forms and
-     * normal vectors.
-     *
-     * This vector has (dim-1)GeometryInfo::faces_per_cell entries. The first
-     * GeometryInfo::faces_per_cell contain the vectors in the first
-     * tangential direction for each face; the second set of
-     * GeometryInfo::faces_per_cell entries contain the vectors in the second
-     * tangential direction (only in 3d, since there we have 2 tangential
-     * directions per face), etc.
-     *
-     * Filled once.
-     */
-    std::vector<std::vector<Tensor<1,dim> > > unit_tangentials;
-
-    /**
-     * The polynomial degree of the mapping. Since the objects here
-     * are also used (with minor adjustments) by MappingQ, we need to
-     * store this.
-     */
-    unsigned int polynomial_degree;
-
-    /**
-     * Number of shape functions. If this is a Q1 mapping, then it is simply
-     * the number of vertices per cell. However, since also derived classes
-     * use this class (e.g. the Mapping_Q() class), the number of shape
-     * functions may also be different.
-     *
-     * In general, it is $(p+1)^\text{dim}$, where $p$ is the
-     * polynomial degree of the mapping.
-     */
-    const unsigned int n_shape_functions;
-
-    /**
-     * Tensors of covariant transformation at each of the quadrature points.
-     * The matrix stored is the Jacobian * G^{-1}, where G = Jacobian^{t} *
-     * Jacobian, is the first fundamental form of the map; if dim=spacedim
-     * then it reduces to the transpose of the inverse of the Jacobian matrix,
-     * which itself is stored in the @p contravariant field of this structure.
-     *
-     * Computed on each cell.
-     */
-    mutable std::vector<DerivativeForm<1,dim, spacedim > >  covariant;
-
-    /**
-     * Tensors of contravariant transformation at each of the quadrature
-     * points. The contravariant matrix is the Jacobian of the transformation,
-     * i.e. $J_{ij}=dx_i/d\hat x_j$.
-     *
-     * Computed on each cell.
-     */
-    mutable std::vector< DerivativeForm<1,dim,spacedim> > contravariant;
-
-    /**
-     * Auxiliary vectors for internal use.
-     */
-    mutable std::vector<std::vector<Tensor<1,spacedim> > > aux;
-
-    /**
-     * Stores the support points of the mapping shape functions on the @p
-     * cell_of_current_support_points.
-     */
-    mutable std::vector<Point<spacedim> > mapping_support_points;
-
-    /**
-     * Stores the cell of which the @p mapping_support_points are stored.
-     */
-    mutable typename Triangulation<dim,spacedim>::cell_iterator cell_of_current_support_points;
-
-    /**
-     * The determinant of the Jacobian in each quadrature point. Filled if
-     * #update_volume_elements.
-     */
-    mutable std::vector<double> volume_elements;
-  };
+  typedef typename MappingQGeneric<dim,spacedim>::InternalData InternalData;
 
 protected:
 
-  // documentation can be found in Mapping::requires_update_flags()
-  virtual
-  UpdateFlags
-  requires_update_flags (const UpdateFlags update_flags) const;
-
-  // documentation can be found in Mapping::get_data()
-  virtual
-  InternalData *
-  get_data (const UpdateFlags,
-            const Quadrature<dim> &quadrature) const;
-
-  // documentation can be found in Mapping::get_face_data()
-  virtual
-  typename Mapping<dim,spacedim>::InternalDataBase *
-  get_face_data (const UpdateFlags flags,
-                 const Quadrature<dim-1>& quadrature) const;
-
-  // documentation can be found in Mapping::get_subface_data()
-  virtual
-  typename Mapping<dim,spacedim>::InternalDataBase *
-  get_subface_data (const UpdateFlags flags,
-                    const Quadrature<dim-1>& quadrature) const;
-
-  // documentation can be found in Mapping::fill_fe_values()
-  virtual
-  CellSimilarity::Similarity
-  fill_fe_values (const typename Triangulation<dim,spacedim>::cell_iterator     &cell,
-                  const CellSimilarity::Similarity                               cell_similarity,
-                  const Quadrature<dim>                                         &quadrature,
-                  const typename Mapping<dim,spacedim>::InternalDataBase        &internal_data,
-                  dealii::internal::FEValues::MappingRelatedData<dim, spacedim> &output_data) const;
-
-  // documentation can be found in Mapping::fill_fe_face_values()
-  virtual void
-  fill_fe_face_values (const typename Triangulation<dim,spacedim>::cell_iterator     &cell,
-                       const unsigned int                                             face_no,
-                       const Quadrature<dim-1>                                       &quadrature,
-                       const typename Mapping<dim,spacedim>::InternalDataBase        &internal_data,
-                       dealii::internal::FEValues::MappingRelatedData<dim, spacedim> &output_data) const;
-
-  // documentation can be found in Mapping::fill_fe_subface_values()
-  virtual void
-  fill_fe_subface_values (const typename Triangulation<dim,spacedim>::cell_iterator     &cell,
-                          const unsigned int                                             face_no,
-                          const unsigned int                                             subface_no,
-                          const Quadrature<dim-1>                                       &quadrature,
-                          const typename Mapping<dim,spacedim>::InternalDataBase        &internal_data,
-                          dealii::internal::FEValues::MappingRelatedData<dim, spacedim> &output_data) const;
 
   /**
    * @}
@@ -478,15 +140,6 @@ protected:
   Point<dim>
   transform_real_to_unit_cell_initial_guess (const std::vector<Point<spacedim> > &vertex,
                                              const Point<spacedim>                            &p) const;
-
-
-  /**
-   * Declare a convenience typedef for the class that describes offsets into
-   * quadrature formulas projected onto faces and subfaces.
-   */
-  typedef
-  typename QProjector<dim>::DataSetDescriptor
-  DataSetDescriptor;
 
 
   /**
@@ -586,154 +239,6 @@ struct StaticMappingQ1
 
 
 /*@}*/
-
-/*----------------------------------------------------------------------*/
-
-#ifndef DOXYGEN
-
-template<int dim, int spacedim>
-inline
-const double &
-MappingQ1<dim,spacedim>::InternalData::shape (const unsigned int qpoint,
-                                              const unsigned int shape_nr) const
-{
-  Assert(qpoint*n_shape_functions + shape_nr < shape_values.size(),
-         ExcIndexRange(qpoint*n_shape_functions + shape_nr, 0,
-                       shape_values.size()));
-  return shape_values [qpoint*n_shape_functions + shape_nr];
-}
-
-
-
-template<int dim, int spacedim>
-inline
-double &
-MappingQ1<dim,spacedim>::InternalData::shape (const unsigned int qpoint,
-                                              const unsigned int shape_nr)
-{
-  Assert(qpoint*n_shape_functions + shape_nr < shape_values.size(),
-         ExcIndexRange(qpoint*n_shape_functions + shape_nr, 0,
-                       shape_values.size()));
-  return shape_values [qpoint*n_shape_functions + shape_nr];
-}
-
-
-template<int dim, int spacedim>
-inline
-const Tensor<1,dim> &
-MappingQ1<dim,spacedim>::InternalData::derivative (const unsigned int qpoint,
-                                                   const unsigned int shape_nr) const
-{
-  Assert(qpoint*n_shape_functions + shape_nr < shape_derivatives.size(),
-         ExcIndexRange(qpoint*n_shape_functions + shape_nr, 0,
-                       shape_derivatives.size()));
-  return shape_derivatives [qpoint*n_shape_functions + shape_nr];
-}
-
-
-
-template<int dim, int spacedim>
-inline
-Tensor<1,dim> &
-MappingQ1<dim,spacedim>::InternalData::derivative (const unsigned int qpoint,
-                                                   const unsigned int shape_nr)
-{
-  Assert(qpoint*n_shape_functions + shape_nr < shape_derivatives.size(),
-         ExcIndexRange(qpoint*n_shape_functions + shape_nr, 0,
-                       shape_derivatives.size()));
-  return shape_derivatives [qpoint*n_shape_functions + shape_nr];
-}
-
-
-template <int dim, int spacedim>
-inline
-const Tensor<2,dim> &
-MappingQ1<dim,spacedim>::InternalData::second_derivative (const unsigned int qpoint,
-                                                          const unsigned int shape_nr) const
-{
-  Assert(qpoint*n_shape_functions + shape_nr < shape_second_derivatives.size(),
-         ExcIndexRange(qpoint*n_shape_functions + shape_nr, 0,
-                       shape_second_derivatives.size()));
-  return shape_second_derivatives [qpoint*n_shape_functions + shape_nr];
-}
-
-
-template <int dim, int spacedim>
-inline
-Tensor<2,dim> &
-MappingQ1<dim,spacedim>::InternalData::second_derivative (const unsigned int qpoint,
-                                                          const unsigned int shape_nr)
-{
-  Assert(qpoint*n_shape_functions + shape_nr < shape_second_derivatives.size(),
-         ExcIndexRange(qpoint*n_shape_functions + shape_nr, 0,
-                       shape_second_derivatives.size()));
-  return shape_second_derivatives [qpoint*n_shape_functions + shape_nr];
-}
-
-template <int dim, int spacedim>
-inline
-const Tensor<3,dim> &
-MappingQ1<dim,spacedim>::InternalData::third_derivative (const unsigned int qpoint,
-                                                         const unsigned int shape_nr) const
-{
-  Assert(qpoint*n_shape_functions + shape_nr < shape_third_derivatives.size(),
-         ExcIndexRange(qpoint*n_shape_functions + shape_nr, 0,
-                       shape_third_derivatives.size()));
-  return shape_third_derivatives [qpoint*n_shape_functions + shape_nr];
-}
-
-
-template <int dim, int spacedim>
-inline
-Tensor<3,dim> &
-MappingQ1<dim,spacedim>::InternalData::third_derivative (const unsigned int qpoint,
-                                                         const unsigned int shape_nr)
-{
-  Assert(qpoint*n_shape_functions + shape_nr < shape_third_derivatives.size(),
-         ExcIndexRange(qpoint*n_shape_functions + shape_nr, 0,
-                       shape_third_derivatives.size()));
-  return shape_third_derivatives [qpoint*n_shape_functions + shape_nr];
-}
-
-
-template <int dim, int spacedim>
-inline
-const Tensor<4,dim> &
-MappingQ1<dim,spacedim>::InternalData::fourth_derivative (const unsigned int qpoint,
-                                                          const unsigned int shape_nr) const
-{
-  Assert(qpoint*n_shape_functions + shape_nr < shape_fourth_derivatives.size(),
-         ExcIndexRange(qpoint*n_shape_functions + shape_nr, 0,
-                       shape_fourth_derivatives.size()));
-  return shape_fourth_derivatives [qpoint*n_shape_functions + shape_nr];
-}
-
-
-template <int dim, int spacedim>
-inline
-Tensor<4,dim> &
-MappingQ1<dim,spacedim>::InternalData::fourth_derivative (const unsigned int qpoint,
-                                                          const unsigned int shape_nr)
-{
-  Assert(qpoint*n_shape_functions + shape_nr < shape_fourth_derivatives.size(),
-         ExcIndexRange(qpoint*n_shape_functions + shape_nr, 0,
-                       shape_fourth_derivatives.size()));
-  return shape_fourth_derivatives [qpoint*n_shape_functions + shape_nr];
-}
-
-
-
-template <int dim, int spacedim>
-inline
-bool
-MappingQ1<dim,spacedim>::preserves_vertex_locations () const
-{
-  return true;
-}
-
-#endif // DOXYGEN
-
-/* -------------- declaration of explicit specializations ------------- */
 
 
 DEAL_II_NAMESPACE_CLOSE
