+++ source/fe/CMakeLists.txt
@@ -52,6 +52,7 @@ SET(_src
   mapping_cartesian.cc
   mapping.cc
   mapping_fe_field.cc
+  mapping_q_generic.cc
   mapping_q1.cc
   mapping_q1_eulerian.cc
   mapping_q.cc
@@ -95,6 +96,7 @@ SET(_inst
   mapping_cartesian.inst.in
   mapping.inst.in
   mapping_fe_field.inst.in
+  mapping_q_generic.inst.in
   mapping_q1_eulerian.inst.in
   mapping_q1.inst.in
   mapping_q_eulerian.inst.in
